requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
rich>=13.0.0
pyyaml>=6.0
flask>=3.0.0
pytest>=8.0.0
numpy>=1.26.0
